draft for multi-label classifier
